refactor: CommonLayout 이동 적용
