+++ src/pages/Home.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import Showcase from '../components/home/showcase/Showcase';
 import AuthPanel from '../components/home/auth_panel/AuthPanel';
 import { CommonHeader } from '../components/common/customHeader';
-import CommonLayout from '../layouts/CommonLayout';
+import CommonLayout from '../components/layouts/CommonLayout';
 
 const carouselItems = [
     {
